api/masto/instance: use bare WebFinger domain for uri
Despite its name this property is not supposed to be a full URI,
but just a bare domain witout protocol. Additionally (and not mentioned
in Masto docs), it’s supposed to be the WebFinger domain used in
userhandles and NOT the domain used for API and ActivityPub objects.

Akkoma’s divergence was reported to
cause troubles in Pachli and Tusky.

Fixes: https://akkoma.dev/AkkomaGang/akkoma/issues/907